+++ src/js/ui/util.js
@@ -70,7 +70,6 @@ export function addLog(msg, type) {
     el.className = `log-type-${type}`;
   }
 
-
   log.appendChild(el);
   // log.innerText += `${msg}\n`;
 
@@ -89,67 +88,4 @@ export function addLog(msg, type) {
 
     log.removeChild(child);
   }, 2500);
-}
-
-
-export function lineBetween(from, to) {
-  if (from === null || to === null) {
-    return false;
-  }
-
-  let id = `line-${from.id}-${to.id}`;
-
-  /*if (document.getElementById(id) !== null) {
-    document.getElementById('sandbox').removeChild(document.getElementById(id));
-  }*/
-
-  [...document.getElementsByClassName('line')].forEach((x) => document.getElementById('sandbox').removeChild(x));
-
-  let line = document.createElement('div');
-  line.className = 'line';
-  line.id = id;
-
-  var fT = from.offsetTop//  + from.offsetHeight/2;
-  var tT = to.offsetTop//    + to.offsetHeight/2;
-  var fL = from.offsetLeft// + from.offsetWidth/2;
-  var tL = to.offsetLeft//   + to.offsetWidth/2;
-  
-  var CA   = Math.abs(tT - fT);
-  var CO   = Math.abs(tL - fL);
-  var H    = Math.sqrt(CA*CA + CO*CO);
-  var ANG  = 180 / Math.PI * Math.acos( CA/H );
-
-  let top;
-  if (tT > fT) {
-    top = (tT-fT)/2 + fT;
-  } else {
-    top = (fT-tT)/2 + tT;
-  }
-
-  let left;
-  if (tL > fL) {
-    left = (tL-fL)/2 + fL;
-  } else {
-    left = (fL-tL)/2 + tL;
-  }
-
-  if ((fT < tT && fL < tL) || (tT < fT && tL < fL) || (fT > tT && fL > tL) || (tT > fT && tL > fL)) {
-    ANG *= -1;
-  }
-
-  top -= H/2;
-
-  //top -= from.offsetHeight / 2;
-  //left -= from.offsetWidth / 2;
-
-  line.style["-webkit-transform"] = 'rotate('+ ANG +'deg)';
-  line.style["-moz-transform"] = 'rotate('+ ANG +'deg)';
-  line.style["-ms-transform"] = 'rotate('+ ANG +'deg)';
-  line.style["-o-transform"] = 'rotate('+ ANG +'deg)';
-  line.style["-transform"] = 'rotate('+ ANG +'deg)';
-  line.style.top = top+'px';
-  line.style.left = left+'px';
-  line.style.height = H + 'px';
-
-  document.getElementById('sandbox').appendChild(line);
 }